refactor site dir
