--- src/applications/mhv/secure-messaging/tests/e2e/keyboard-nav-test/secure-messaging-keyboard-nav-to-compose.cypress.spec.js
@@ -1,23 +1,0 @@
-import SecureMessagingWelcomePage from '../pages/SecureMessagingWelcomePage';
-import SecureMessagingSite from '../sm_site/SecureMessagingSite';
-
-describe('Secure Messaging Compose', () => {
-  const landingPage = new SecureMessagingWelcomePage();
-  const site = new SecureMessagingSite();
-  beforeEach(() => {
-    site.login();
-    landingPage.loadPage();
-  });
-  it('Tab to Message Body', () => {
-    cy.tabToElement('[data-testid="compose-message-link"]').pushEnter();
-    cy.injectAxe();
-    cy.axeCheck();
-    cy.tabToElement('[data-testid="message-body-field"] ').should('exist');
-  });
-  it('Tab to Save Draft Button', () => {
-    cy.get('[data-testid="compose-message-link"]').click();
-    cy.injectAxe();
-    cy.axeCheck();
-    cy.tabToElement('[data-testid="Save-Draft-Button"]').should('exist');
-  });
-});
